fix: Entity, NPC Move, SkillBuffer
